requirements.txt
modal-client==0.51.3192
cohere==4.21
python-dotenv==1.0.0

fastapi==0.102.0
pydantic
pymongo
slack_sdk==3.21.3
jinja2
uvicorn
asgiref
numpy
anthropic==0.2.9
motor

# Local
black
isort
